+++ R/expr__list.R
@@ -207,6 +207,7 @@ ExprList_contains = function(item) .pr$Expr$list_contains(self, wrap_e(item))
 #'
 #' @param separator String to separate the items with. Can be an Expr. Strings
 #'   are *not* parsed as columns.
+#' @inheritParams pl_concat_str
 #'
 #' @return Expr
 #' @aliases list_join
@@ -217,10 +218,11 @@ ExprList_contains = function(item) .pr$Expr$list_contains(self, wrap_e(item))
 #' )
 #' df$with_columns(
 #'   join_with_expr = pl$col("s")$list$join(pl$col("separator")),
-#'   join_with_lit = pl$col("s")$list$join(" ")
+#'   join_with_lit = pl$col("s")$list$join(" "),
+#'   join_ignore_null = pl$col("s")$list$join(" ", ignore_nulls = TRUE)
 #' )
-ExprList_join = function(separator) {
-  .pr$Expr$list_join(self, separator) |>
+ExprList_join = function(separator, ignore_nulls = FALSE) {
+  .pr$Expr$list_join(self, separator, ignore_nulls) |>
     unwrap("in $list$join():")
 }