+++ src/rust/src/arrow_interop/to_rust.rs
@@ -158,7 +158,7 @@ fn consume_arrow_stream_to_series(boxed_stream: Box<ffi::ArrowArrayStream>) -> R
 pub unsafe fn export_df_as_stream(df: pl::DataFrame, robj_str_ref: &Robj) -> RResult<()> {
     let stream_ptr =
         crate::utils::robj_str_ptr_to_usize(robj_str_ref)? as *mut ffi::ArrowArrayStream;
-    let schema = df.schema().to_arrow();
+    let schema = df.schema().to_arrow(true);
     let data_type = pl::ArrowDataType::Struct(schema.fields);
     let field = pl::ArrowField::new("", data_type, false);
     let iter_boxed = Box::new(crate::rdataframe::OwnedDataFrameIterator::new(df));