+++ src/rust/src/rdataframe/mod.rs
@@ -37,7 +37,7 @@ pub struct OwnedDataFrameIterator {
 
 impl OwnedDataFrameIterator {
     pub fn new(df: polars::frame::DataFrame) -> Self {
-        let schema = df.schema().to_arrow();
+        let schema = df.schema().to_arrow(false);
         let data_type = ArrowDataType::Struct(schema.fields);
         let vs = df.get_columns().to_vec();
         Self {
@@ -57,7 +57,11 @@ impl Iterator for OwnedDataFrameIterator {
             None
         } else {
             // create a batch of the columns with the same chunk no.
-            let batch_cols = self.columns.iter().map(|s| s.to_arrow(self.idx)).collect();
+            let batch_cols = self
+                .columns
+                .iter()
+                .map(|s| s.to_arrow(self.idx, false))
+                .collect();
             self.idx += 1;
 
             let chunk = polars::frame::ArrowChunk::new(batch_cols);
@@ -150,11 +154,11 @@ impl RPolarsDataFrame {
             .map_err(|err| format!("in set_column_from_series: {:?}", err))
     }
 
-    pub fn with_row_count(&self, name: Robj, offset: Robj) -> RResult<Self> {
+    pub fn with_row_index(&self, name: Robj, offset: Robj) -> RResult<Self> {
         Ok(self
             .0
             .clone()
-            .with_row_count(
+            .with_row_index(
                 robj_to!(String, name)?.as_str(),
                 robj_to!(Option, u32, offset)?,
             )
@@ -325,7 +329,7 @@ impl RPolarsDataFrame {
     }
 
     pub fn export_stream(&self, stream_ptr: &str) {
-        let schema = self.0.schema().to_arrow();
+        let schema = self.0.schema().to_arrow(false);
         let data_type = ArrowDataType::Struct(schema.fields);
         let field = ArrowField::new("", data_type, false);
 
@@ -480,7 +484,7 @@ impl RPolarsDataFrame {
             .with_separator(robj_to!(Utf8Byte, separator)?)
             .with_line_terminator(robj_to!(String, line_terminator)?)
             .with_quote_char(robj_to!(Utf8Byte, quote)?)
-            .with_batch_size(robj_to!(usize, batch_size)?)
+            .with_batch_size(robj_to!(nonzero_usize, batch_size)?)
             .with_datetime_format(robj_to!(Option, String, datetime_format)?)
             .with_date_format(robj_to!(Option, String, date_format)?)
             .with_time_format(robj_to!(Option, String, time_format)?)