+++ src/rust/src/rdataframe/read_ndjson.rs
@@ -3,7 +3,7 @@
 use crate::lazy::dataframe::RPolarsLazyFrame;
 use crate::robj_to;
 use crate::rpolarserr::*;
-use polars::io::RowCount;
+use polars::io::RowIndex;
 
 //use crate::utils::wrappers::*;
 use extendr_api::{extendr, prelude::*, Rinternals};
@@ -20,12 +20,13 @@ pub fn new_from_ndjson(
     n_rows: Robj,
     low_memory: Robj,
     rechunk: Robj,
-    row_count_name: Robj,
-    row_count_offset: Robj,
+    row_index_name: Robj,
+    row_index_offset: Robj,
+    ignore_errors: Robj,
 ) -> RResult<RPolarsLazyFrame> {
-    let offset = robj_to!(Option, u32, row_count_offset)?.unwrap_or(0);
-    let opt_rowcount =
-        robj_to!(Option, String, row_count_name)?.map(|name| RowCount { name, offset });
+    let offset = robj_to!(Option, u32, row_index_offset)?.unwrap_or(0);
+    let opt_rowindex =
+        robj_to!(Option, String, row_index_name)?.map(|name| RowIndex { name, offset });
 
     let vec_pathbuf = robj_to!(Vec, PathBuf, path)?;
     let linereader = match vec_pathbuf.len() {
@@ -36,11 +37,12 @@ pub fn new_from_ndjson(
 
     linereader
         .with_infer_schema_length(robj_to!(Option, usize, infer_schema_length)?)
-        .with_batch_size(robj_to!(Option, usize, batch_size)?)
+        .with_batch_size(robj_to!(Option, nonzero_usize, batch_size)?)
         .with_n_rows(robj_to!(Option, usize, n_rows)?)
         .low_memory(robj_to!(bool, low_memory)?)
-        .with_row_count(opt_rowcount)
+        .with_row_index(opt_rowindex)
         .with_rechunk(robj_to!(bool, rechunk)?)
+        .with_ignore_errors(robj_to!(bool, ignore_errors)?)
         .finish()
         .map_err(polars_to_rpolars_err)
         .map(RPolarsLazyFrame)