+++ tests/testthat/test-parquet.R
@@ -19,7 +19,7 @@ test_that("plain scan read parquet", {
 })
 
 
-test_that("scan read parquet - test arg rowcount", {
+test_that("scan read parquet - test arg row_index", {
   tmpf = tempfile()
   on.exit(unlink(tmpf))
   lf_exp = pl$LazyFrame(mtcars)
@@ -27,12 +27,12 @@ test_that("scan read parquet - test arg rowcount", {
   df_exp = lf_exp$collect()$to_data_frame()
 
   expect_identical(
-    pl$scan_parquet(tmpf, row_count_name = "rc", row_count_offset = 5)$collect()$to_data_frame(),
+    pl$scan_parquet(tmpf, row_index_name = "rc", row_index_offset = 5)$collect()$to_data_frame(),
     data.frame(rc = as.numeric(5:36), df_exp)
   )
 
   expect_identical(
-    pl$read_parquet(tmpf, row_count_name = "rc", row_count_offset = 5)$to_data_frame(),
+    pl$read_parquet(tmpf, row_index_name = "rc", row_index_offset = 5)$to_data_frame(),
     data.frame(rc = as.numeric(5:36), df_exp)
   )
 })